+++ src/vsmtp/vsmtp-rule-engine/src/dsl/service/smtp/mod.rs
@@ -14,10 +14,5 @@
  * this program. If not, see https://www.gnu.org/licenses/.
  *
 */
-mod aliases;
-mod anti_relaying;
-mod dnsbl;
-mod family;
-mod greylist;
-mod message;
-mod services;
+
+pub mod parsing;